check vid forma z

+++ Fond/src/main/java/by/belisa/entity/VidFormaZ.java
@@ -23,13 +23,14 @@ public class VidFormaZ implements Serializable{
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator="PK")
 	@Column(name="id")
-	private long id;
+	private Integer id;
 	@Column(name="name")
 	private String name;
-	public long getId() {
+		
+	public Integer getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getName() {
